Allow contextualisation of sub-types.
The previous implementation only allowed the explicit type parameter to be contextualised - e.g. the case class typeclass only contextualised Products, meaning that they had to be cast or have their type explicitly set to Product rather than inferred, which is not useful in a strongly typed language.

Making the type parameter contravariant solved this.