Start an alternative implementation of the renderer.
Takes a strongly typed Template made up of Components responsible for rendering themselves. This opens up the possibility of a more efficient renderer which can cache parsed templates rather than reparsing the template every time it is rendered as per the previous implementation.

This is a Work In Progress so the acceptance and integration tests have been marked as pending.