Start to implement a Set Delimiters tag.
Implemented as a "parser directive", a new type of component which does not render, which can be used to modify the parser config on-the-fly.

The parser config now specifies the delimiters used to parse; as yet no other tag uses these delimiters when parsing and so this has no real effect.